Fix transpose() for dynamic size arrays
Before this change, something like transpose(rank2A, {0, 0}) wasn't
possible because the rank of the result was assumed to be the same as
the rank of the argument.

Now transpose(A, {...}) will always return a dynamic rank array, and
transpose<...>(A) or transpose(A, mp::int_list<...>) will always
return a static rank array, even when the argument has dynamic rank.

* ra/tuple-list.H (Iota): Handle step != 1.

* test/test-tuplelist.C: test Iota with step != 1.

* ra/ra-large.H (transpose): new transpose<...>(). transpose(, {})
  always returns dynamic rank.

* ra/ra-operators.H (transpose(, mp::int_list<>)): a convenience.

* test/test-ra-0.C, test/test-ra-1.C: some tests with the new
  transpose() versions.

* test/test-ra-io.C: fix test for different rank type return from
  transpose().